+++ crates/reqsign-google/src/credential/external_account.rs
