+++ crates/reqsign-google/src/credential/impersonated_service_account.rs
