+++ crates/reqsign-google/testdata/test_impersonated_service_account.json
@@ -1,0 +1,11 @@
+{
+    "delegates": [],
+    "service_account_impersonation_url": "https://iamcredentials.googleapis.com/v1/projects/-/serviceAccounts/example-01-iam@example-01.iam.gserviceaccount.com:generateAccessToken",
+    "source_credentials": {
+        "client_id": "placeholder_client_id",
+        "client_secret": "placeholder_client_secret",
+        "refresh_token": "placeholder_refresh_token",
+        "type": "authorized_user"
+    },
+    "type": "impersonated_service_account"
+}
