+++ crates/reqsign-google/testdata/testbucket_credential.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "iam-testbucket-reqsign-project",
+  "private_key_id": "abcdefabcdef4f96c0c0ed6cffe88a6413e91cec",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDMXQUceY0V5Id3\nhT4Xu/eUOkj7vYMxmcM7u03g/R5d1Jj+5WdpeLu/Kjh2EhXuCfodCSawEoU1O66h\nfL9A2EMefHRE3tPMYDCYJgUqes/6of4QcSJKev29blbryKrNfQEp4xzRX9oY68bs\nRNVoDN/tIKovZle7VvOPjK83HeTv/R9l93dbKS1lTIYrNI86aRhvvnfbDZGTq/iI\nMHreTD8ICc06qsH774uVMBhTH3bxRQ8NTrZP70PS8RfCCTVNwiwSPE146aU6XUWU\nh+29rqbMNHHQsWLCLeLTY565kD+a2DtNIU0H89wPxnirDouZBYwCjtivvlE0mN5y\nj1qkPrwfAgMBAAECggEBAMO6k5qiEC5XoicmxkGVFZox+JSi/XQUAJjE2+IQi3Ty\nmVYIAPNTXv3IQitTRw2lIJeOnC8mjc5eSvL/t20zs5UPPYx4ngGwXtpaD7iPx4IU\nhHDa6izLfxpfA4DvwCbvAp5Llt4xH4Gez/aaNophSlaiYlzjeENFFCD4bRgs2Ye+\n/pyF0akNf7RO2PXC0u0KbP3rf/tqY+tGnNg7Ykx3+4yEetwFW2RgL3IlLU4PM2Xg\nRrjdiLHNOjS1VXCBMaEYuoP1HFCR+JzyBdidD++/kTSHVCab9Dz4HemyKxr8gpYK\nZUREirN8GFZwVt+SysWzIh788adhhy7Tscyy6SoQ2AECgYEA+76McDcxlEGTkRuO\nXh/dEeFDnSGEqOC3HfYJbzlo+K5bD9ged3DtwSN+bxDizHtcQnhqMPOhrD0tgiCn\nmNsEPXAhj3eoRd050R6THEXwE5Az84SR9gvm27Q8JoAUWXWIkrpVKPwDDjq7XoIF\nY1lTld84UGwg8Hgux05KcMZROd8CgYEAz9FszBaX1OgmRcOklxmDgFsWsdiepFpL\nrTj6oNpbHiJYYDJ0DQgvty8dI+1aWfqgHh1TczjYRx3Cf77qUGpYzY/Vkc2lC+fU\naxXz3qmaKIYZH/IA5+EZshXjMSwTTaGzpmzoqXZ57J9uwmQeUsLeXnlZd1habk7O\nD3CZhuK3RcECgYBGWUdRjHr0XSbpo/Oy5eCXQIXugRFbSACkBL86L6bf54lW8iQB\naLNoB40raGKYldiAUroKF+sUALyY4pszIfEbYhxexSdm7p1bjNm7Suf974w0/tTz\nFvxaZRFyCNSm8ytJJXzqyRHphgwaKudqjenHtes8vhquWEdqNryiqyjDrQKBgCV7\nNBArEv9HT3/NpWXLKDiCNTmmRBaIYpW/bRSNzVlGAIJ5Fw0yqMh1KuBL8ru/xBkq\nWN6zJe7No0K/ACu4woNwqag+WsIm8dzOfMlv9WnRpb5pO1iW9Ld10yAPPvwFag1e\nHyhRQfQ3XRaaUA3FL64CXOx1dvnmJKwMNuRpB30BAoGBAJ+DhBf9CM6ndMBJaOZa\nD8Let93+NYu/iJEn2whzN1cGBcnsOw9uAWiOkFUOuNqYVzSih11vLvUIw9Q9LKnI\n8MLsFLjvierQ1h+V1mxhNO8bEv9tTO+O8h1GMNmCy4S0wZjfsbrJzPL+pKrpgEJS\nqjDDLD9IP7jvoKHSODDsecNq\n-----END PRIVATE KEY-----\n",
+  "client_email": "testbucket-reqsign-account@iam-testbucket-reqsign-project.iam.gserviceaccount.com",
+  "client_id": "101000000000000000000",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/testbucket-reqsign-account%iam-testbucket-reqsign-project.iam.gserviceaccount.iam.gserviceaccount.com"
+}
