+++ crates/reqsign/Cargo.toml
@@ -14,20 +14,7 @@ all-features = true
 
 [features]
 # services that reqsign supports
-services-all = [
-  "services-google",
-  "services-huaweicloud",
-  "services-oracle",
-  "services-tencent",
-]
-
-services-google = [
-  "dep:reqwest",
-  "dep:serde",
-  "dep:serde_json",
-  "dep:jsonwebtoken",
-  "dep:rsa",
-]
+services-all = ["services-huaweicloud", "services-oracle", "services-tencent"]
 services-huaweicloud = ["dep:serde", "dep:serde_json", "dep:once_cell"]
 services-oracle = ["dep:reqwest", "dep:rsa", "dep:toml", "dep:serde"]
 services-tencent = ["dep:reqwest", "dep:serde", "dep:serde_json"]
@@ -41,11 +28,9 @@ form_urlencoded.workspace = true
 hex.workspace = true
 hmac.workspace = true
 http.workspace = true
-jsonwebtoken = { version = "9.2", optional = true }
 log.workspace = true
 once_cell = { workspace = true, optional = true }
 percent-encoding.workspace = true
-rand.workspace = true
 reqwest = { workspace = true, optional = true }
 rsa = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
