--- crates/reqsign/tests/google/mod.rs
@@ -1,1 +1,0 @@
-mod storage;
