+++ crates/reqsign/tests/main.rs
@@ -1,2 +1,1 @@
-mod google;
 mod tencent;
